move identify from ZendeskSupportComponent to ZendeskAskPermission screen
